requirements.txt
llama_index==0.10.30
protobuf==3.20.3
python-dotenv==1.0.1
streamlit==1.30.0
streamlit==1.32.2
